Fix and extend SysEx API
This commit fixes the existing WriteSysEx function to allow writing a
hex string to the output stream.

It also introduces WriteSysExBytes() to provide the message as a byte
slice instead.

On the input side, ReadSysExBytes() reads all four bytes of up to max
PmMessages and returns them as a byte slice.